feat: enhance Azure Foundry Agents plugin with improved error handling and polling mechanism
- Updated the Azure Foundry Agents plugin to utilize axios for HTTP requests, improving reliability and readability.
- Enhanced error handling to suggest using a backup tool when the primary tool fails.
- Refined polling logic for run completion, removing deprecated request methods and improving response handling.
- Added tests to verify access to the azureAuthTokenHelper and handle cases where it is missing.
