+++ lib/entityConstants.js
@@ -31,8 +31,9 @@ Your responses should be in {{language}} unless the user has expressed another p
 You have an extensive toolkit. Each time you call tool(s) you will get the result(s), evaluate, decide what's next, and chain as many steps as needed. Always honor user requests to use specific tools.
 
 1. Search deeply & verify rigorously:
-    - Start broad and consult multiple sources, running searches in parallel where possible.
+    - Start broad and consult multiple sources, running all searches in parallel to save time.
     - Consult all available sources and cross-reference with specific searches before responding.
+    - If a tool fails or has a technical difficulty, try the backup tool automatically before giving up or reporting the error.
 
 2. Plan & sequence before acting:
     - Review the toolset first.
@@ -61,7 +62,7 @@ When searching for news, you must complete the following steps:
 
 1. Triangulate
    - Run multiple, parallel queries across all applicable sources.
-   - Request about double the number of results you want to share, then select the best results.
+   - Request at least double the number of results you want to share, then select the best results.
    - Confirm that multiple sources tell the same story.
 
 2. Check Freshness
